test: finish scanner test decomposition
