+++ tests/test_scanner_coverage.py
@@ -371,19 +371,6 @@ async def test_is_valid_bcd_time_valid_value():
 
 
 @pytest.mark.asyncio
-async def test_safe_scan_group_registers():
-    """Line 994: safe_scan=True → single-register batches."""
-    scanner = await _make_scanner(safe_scan=True)
-    result = scanner._group_registers_for_batch_read([0, 1, 5, 10])
-    assert result == [(0, 1), (1, 1), (5, 1), (10, 1)]
-
-
-# ---------------------------------------------------------------------------
-# Group K: scan() raises ConnectionException without transport/client
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
 async def test_scan_raises_without_transport_and_client():
     """Lines 1027-1028: transport=None and client=None raises ConnectionException."""
     scanner = await _make_scanner()
@@ -1039,279 +1026,6 @@ async def _run_minimal_scan(
 
 
 @pytest.mark.asyncio
-async def test_scan_skip_known_missing_input_register():
-    """Line 1302-1303: skip_known_missing=True skips 'compilation_days'."""
-    scanner = await _make_scanner(skip_known_missing=True)
-    scanner._client = AsyncMock()
-
-    from custom_components.thessla_green_modbus.scanner.core import INPUT_REGISTERS
-
-    if "compilation_days" not in INPUT_REGISTERS:
-        pytest.skip("compilation_days not in INPUT_REGISTERS")
-
-    result = await _run_minimal_scan(scanner, input_return=[1])
-    # compilation_days should not appear in missing registers (was skipped)
-    missing = result.get("missing_registers", {}).get("input_registers", {})
-    assert "compilation_days" not in missing
-
-
-@pytest.mark.asyncio
-async def test_scan_input_batch_fail_probe_success():
-    """Lines 1313-1341: batch read fails, probe individual succeeds."""
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-
-    from custom_components.thessla_green_modbus.scanner.core import INPUT_REGISTERS
-
-    if "version_major" not in INPUT_REGISTERS:
-        pytest.skip("version_major not in INPUT_REGISTERS")
-
-    addr = INPUT_REGISTERS["version_major"]
-    scanner._registers = {4: {addr: "version_major"}, 3: {}, 1: {}, 2: {}}
-
-    batch_call_count = {"n": 0}
-
-    async def mock_read_input(*args, **kwargs):
-        batch_call_count["n"] += 1
-        if batch_call_count["n"] <= 2:  # firmware block reads return empty
-            return []
-        count = args[-1] if len(args) > 1 else kwargs.get("count", 1)
-        if count > 1:
-            return None  # batch fails
-        return [4]  # individual probe succeeds
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", side_effect=mock_read_input),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert "available_registers" in result
-
-
-@pytest.mark.asyncio
-async def test_scan_input_batch_fail_probe_fail(caplog):
-    """Lines 1332-1334: batch fails, individual probe returns falsy → warning."""
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-
-    from custom_components.thessla_green_modbus.scanner.core import INPUT_REGISTERS
-
-    if "version_major" not in INPUT_REGISTERS:
-        pytest.skip("version_major not in INPUT_REGISTERS")
-
-    addr = INPUT_REGISTERS["version_major"]
-    scanner._registers = {4: {addr: "version_major"}, 3: {}, 1: {}, 2: {}}
-
-    async def mock_read_input(*args, **kwargs):
-        count = args[-1] if len(args) > 1 else 1
-        if isinstance(count, int) and count > 1:
-            return None  # batch fails
-        return None  # probe also fails
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", side_effect=mock_read_input),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-        caplog.at_level(logging.WARNING),
-    ):
-        await scanner.scan()
-
-    assert "Failed to read input_registers register" in caplog.text
-
-
-# ---------------------------------------------------------------------------
-# Group P: Holding batch failure recovery (lines 1368-1410)
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_scan_holding_batch_fail_probe_success():
-    """Lines 1373-1400: holding batch fails, probe succeeds."""
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-
-    from custom_components.thessla_green_modbus.scanner.core import HOLDING_REGISTERS
-
-    if "mode" not in HOLDING_REGISTERS:
-        pytest.skip("mode not in HOLDING_REGISTERS")
-
-    addr = HOLDING_REGISTERS["mode"]
-    scanner._registers = {4: {}, 3: {addr: "mode"}, 1: {}, 2: {}}
-
-    async def mock_read_holding(*args, **kwargs):
-        count = args[-1] if len(args) > 1 else 1
-        if isinstance(count, int) and count > 1:
-            return None  # batch fails
-        return [1]  # probe succeeds with valid value
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", side_effect=mock_read_holding),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert "available_registers" in result
-
-
-# ---------------------------------------------------------------------------
-# Group Q: deep_scan=True (line 1548)
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_deep_scan_skips_none_result():
-    """Line 1548: deep_scan=True with None read results → continue."""
-    scanner = await _make_scanner(deep_scan=True)
-    scanner._client = AsyncMock()
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert "raw_registers" in result
-    assert result["raw_registers"] == {}
-
-
-@pytest.mark.asyncio
-async def test_deep_scan_collects_values():
-    """Line 1547-1550: deep_scan=True with data collects raw_registers."""
-    scanner = await _make_scanner(deep_scan=True)
-    scanner._client = AsyncMock()
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=[42])),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert "raw_registers" in result
-    assert len(result["raw_registers"]) > 0
-
-
-# ---------------------------------------------------------------------------
-# Group M: full_register_scan with invalid holding values (lines 1229-1253)
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_input_returns_none():
-    """Lines 1198-1202: full_register_scan input read returns None."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {0: "version_major"}, 3: {0: "mode"}, 1: {}, 2: {}}
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert result["failed_addresses"]["modbus_exceptions"]["input_registers"]
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_holding_invalid_value():
-    """Lines 1248-1253: full_register_scan holding has invalid value (65535)."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {0: "version_major"}, 3: {0: "mode"}, 1: {}, 2: {}}
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=[1])),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=[65535])),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    # 65535 is invalid → should appear in failed_addresses
-    assert (
-        result["failed_addresses"]["invalid_values"].get("holding_registers")
-        or result["unknown_registers"]["holding_registers"]
-    )
-
-
-# ---------------------------------------------------------------------------
-# Group N: full_register_scan coil/discrete (lines 1258-1296)
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_coil_returns_none():
-    """Lines 1261-1265: full_register_scan coil read returns None."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {0: "version_major"}, 3: {}, 1: {0: "some_coil"}, 2: {}}
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=[1])),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert result["failed_addresses"]["modbus_exceptions"]["coil_registers"]
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_discrete_returns_value():
-    """Lines 1280-1296: full_register_scan discrete reads a value."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {}, 3: {}, 1: {}, 2: {0: "some_discrete"}}
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=[True])),
-    ):
-        result = await scanner.scan()
-
-    # discrete read returned [True] — the register at addr 0 should be in
-    # available_registers (possibly under the global alias name)
-    assert result["available_registers"]["discrete_inputs"]
-
-
-# ---------------------------------------------------------------------------
-# Group S: RTU in scan_device (lines 1669-1675)
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
 async def test_scan_device_rtu_no_serial_port_raises():
     """Line 1669-1670: scan_device with RTU and no serial_port raises."""
     scanner = await _make_scanner(connection_type=CONNECTION_TYPE_RTU)
@@ -1477,447 +1191,6 @@ def _sized_read_mock(value=1):
         return [value] * count
 
     return _mock
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_input_no_alias_path():
-    """Line 1217: input register not in _names_by_address → add by reg_name."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    # Use addr 0 which IS in global but we clear _names_by_address
-    scanner._registers = {4: {0: "fake_input_reg"}, 3: {}, 1: {}, 2: {}}
-    scanner._names_by_address[4] = {}  # empty → _alias_names returns empty set
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", side_effect=_sized_read_mock(1)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    # fake_input_reg should be in available_registers (added without alias)
-    assert "fake_input_reg" in result["available_registers"]["input_registers"]
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_input_invalid_value():
-    """Lines 1221-1222: input register returns invalid value (65535)."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {0: "fake_input_reg"}, 3: {}, 1: {}, 2: {}}
-    scanner._names_by_address[4] = {}
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", side_effect=_sized_read_mock(65535)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    # 65535 is invalid → should be in invalid_values
-    assert 0 in result["failed_addresses"]["invalid_values"]["input_registers"]
-
-
-# ---------------------------------------------------------------------------
-# Line 1248: full_register_scan holding no-alias path
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_holding_no_alias_path():
-    """Line 1248: holding register not in _names_by_address → add by reg_name."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {}, 3: {0: "fake_holding_reg"}, 1: {}, 2: {}}
-    scanner._names_by_address[3] = {}  # empty → no alias
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", side_effect=_sized_read_mock(1)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert "fake_holding_reg" in result["available_registers"]["holding_registers"]
-
-
-# ---------------------------------------------------------------------------
-# Lines 1266-1275: full_register_scan coil valid response
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_coil_valid_with_name():
-    """Lines 1266-1275: full_register_scan coil returns data, register in map."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {}, 3: {}, 1: {0: "fake_coil"}, 2: {}}
-    scanner._names_by_address[1] = {}  # no alias
-
-    async def coil_mock(*args, **kw):
-        count = 1
-        for a in reversed(args):
-            if isinstance(a, int):
-                count = a
-                break
-        return [True] * count
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", side_effect=coil_mock),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert "fake_coil" in result["available_registers"]["coil_registers"]
-
-
-# ---------------------------------------------------------------------------
-# Lines 1283-1286, 1294-1296: full_register_scan discrete None and valid
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_discrete_none():
-    """Lines 1283-1286: full_register_scan discrete returns None."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {}, 3: {}, 1: {}, 2: {0: "fake_discrete"}}
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert 0 in result["failed_addresses"]["modbus_exceptions"]["discrete_inputs"]
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_discrete_valid():
-    """Lines 1294-1296: full_register_scan discrete returns valid data."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {}, 3: {}, 1: {}, 2: {0: "fake_discrete"}}
-    scanner._names_by_address[2] = {}  # no alias
-
-    async def discrete_mock(*args, **kw):
-        count = 1
-        for a in reversed(args):
-            if isinstance(a, int):
-                count = a
-                break
-        return [True] * count
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", side_effect=discrete_mock),
-    ):
-        result = await scanner.scan()
-
-    assert "fake_discrete" in result["available_registers"]["discrete_inputs"]
-
-
-# ---------------------------------------------------------------------------
-# Line 1325: input probe fails → warning
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_scan_input_probe_always_fails(caplog):
-    """Line 1325: batch fails, individual probe returns falsy → warning."""
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {9999: "fake_input"}, 3: {}, 1: {}, 2: {}}
-    scanner._names_by_address[4] = {9999: {"fake_input"}}
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-        caplog.at_level(logging.WARNING),
-    ):
-        await scanner.scan()
-
-    assert "Failed to read input_registers register 9999" in caplog.text
-
-
-# ---------------------------------------------------------------------------
-# Lines 1339-1340: input probe returns invalid value
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_scan_input_probe_returns_invalid_value():
-    """Lines 1339-1340: batch fails, probe returns invalid value (65535)."""
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {9999: "fake_input"}, 3: {}, 1: {}, 2: {}}
-    scanner._names_by_address[4] = {9999: {"fake_input"}}
-
-    call_n = {"n": 0}
-
-    async def mock_read_input(*args, **kwargs):
-        call_n["n"] += 1
-        # First 2 calls are _read_input_block's firmware read (returns [])
-        # but we patch _read_input_block separately
-        # The direct _read_input calls: batch (count=1 for addr 9999) → None
-        # then probe → [65535]
-        count = args[-1] if len(args) >= 2 else kwargs.get("count", 1)
-        if count is None:
-            count = 1
-        return None  # all fail
-
-    # Actually, simplest: batch=None, probe=[65535]
-    batch_n = {"n": 0}
-
-    async def read_input_for_probe(*args, **kwargs):
-        batch_n["n"] += 1
-        if batch_n["n"] == 1:
-            return None  # batch fails
-        return [65535]  # probe returns invalid
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", side_effect=read_input_for_probe),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert 9999 in result["failed_addresses"]["invalid_values"]["input_registers"]
-
-
-# ---------------------------------------------------------------------------
-# Line 1357: holding scan skips UART-optional registers
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_scan_holding_skips_uart_optional_registers():
-    """Line 1357: scan_uart_settings=False skips UART optional registers."""
-    scanner = await _make_scanner(scan_uart_settings=False, retry=1)
-    scanner._client = AsyncMock()
-    # UART_OPTIONAL_REGS = range(4452, 4460)
-    scanner._registers = {4: {}, 3: {4452: "uart_0_id"}, 1: {}, 2: {}}
-
-    read_holding_calls = []
-
-    async def mock_read_holding(*args, **kwargs):
-        read_holding_calls.append(args)
-        return [1]
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", side_effect=mock_read_holding),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        await scanner.scan()
-
-    # addr 4452 should have been skipped — not read
-    assert not read_holding_calls
-
-
-# ---------------------------------------------------------------------------
-# Lines 1362-1363: holding multi-register (MULTI_REGISTER_SIZES > 1)
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_scan_holding_multiregister_alias():
-    """Lines 1362-1363: duplicate holding address merges names into one entry."""
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-
-    class DuplicateAddressHoldingMap(dict[int, str]):
-        def items(self):
-            return iter([(9999, "fake_h1"), (9999, "fake_h2")])
-
-    with (
-        patch.object(scanner, "_group_registers_for_batch_read", return_value=[(9999, 1)]),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=[123])),
-        patch.object(scanner, "_is_valid_register_value", return_value=True),
-    ):
-        await scanner._scan_named_holding(DuplicateAddressHoldingMap())
-
-    assert "fake_h1" in scanner.available_registers["holding_registers"]
-    assert "fake_h2" in scanner.available_registers["holding_registers"]
-
-
-# ---------------------------------------------------------------------------
-# Lines 1371-1372: holding TypeError fallback in batch read
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_scan_holding_type_error_fallback():
-    """Lines 1371-1372: _read_holding raises TypeError → fallback to 2-arg."""
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {}, 3: {9999: "fake_holding"}, 1: {}, 2: {}}
-    scanner._names_by_address[3] = {9999: {"fake_holding"}}
-
-    call_n = {"n": 0}
-
-    async def mock_read_holding(*args, **kw):
-        call_n["n"] += 1
-        if len(args) == 3 and call_n["n"] == 1:
-            raise TypeError("unexpected signature")
-        return [1]  # fallback (2-arg) succeeds
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", side_effect=mock_read_holding),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert "fake_holding" in result["available_registers"]["holding_registers"]
-
-
-# ---------------------------------------------------------------------------
-# Lines 1384, 1398-1399: holding probe addr not in map / invalid probe value
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_scan_holding_probe_addr_not_in_info():
-    """Line 1384: addr in batch range but not in holding_info → continue."""
-    # To get a batch range wider than holding_info, we need _group_reads to
-    # merge two addresses with a gap between them — but it shouldn't unless
-    # the gap is ≤ max_gap. Let's instead trigger it via MULTI_REGISTER_SIZES:
-    # a holding reg with size=2 adds both addr and addr+1 to holding_addresses.
-    # If only the base addr is in holding_info, then addr+1 triggers line 1384.
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-
-    import custom_components.thessla_green_modbus.scanner.core as sc
-
-    original_multi = dict(sc.MULTI_REGISTER_SIZES)
-    try:
-        # Register "fake_multi" at addr 9999 with size 2
-        sc.MULTI_REGISTER_SIZES["fake_multi"] = 2
-        scanner._registers = {4: {}, 3: {9999: "fake_multi"}, 1: {}, 2: {}}
-        scanner._names_by_address[3] = {9999: {"fake_multi"}}
-
-        # batch will cover (9999, 2) → when it fails, we iterate 9999 and 10000
-        # 10000 is not in holding_info → line 1384 hit
-
-        with (
-            patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-            patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-            patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-            patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-            patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-            patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-        ):
-            await scanner.scan()
-
-    finally:
-        sc.MULTI_REGISTER_SIZES.clear()
-        sc.MULTI_REGISTER_SIZES.update(original_multi)
-
-
-@pytest.mark.asyncio
-async def test_scan_holding_probe_invalid_value():
-    """Lines 1398-1399: batch fails, probe returns invalid value → tracking."""
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {}, 3: {9999: "fake_holding"}, 1: {}, 2: {}}
-    scanner._names_by_address[3] = {9999: {"fake_holding"}}
-
-    batch_n = {"n": 0}
-
-    async def read_holding_for_probe(*args, **kwargs):
-        batch_n["n"] += 1
-        if batch_n["n"] == 1:
-            return None  # batch fails
-        return [65535]  # probe returns invalid
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", side_effect=read_holding_for_probe),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert 9999 in result["failed_addresses"]["invalid_values"]["holding_registers"]
-
-
-# ---------------------------------------------------------------------------
-# Lines 1389-1390: TypeError in holding probe → fallback
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_scan_holding_probe_type_error_fallback():
-    """Lines 1389-1390: individual probe raises TypeError → 2-arg fallback."""
-    scanner = await _make_scanner(retry=1)
-    scanner._client = AsyncMock()
-    scanner._registers = {4: {}, 3: {9999: "fake_holding"}, 1: {}, 2: {}}
-    scanner._names_by_address[3] = {9999: {"fake_holding"}}
-
-    call_n = {"n": 0}
-
-    async def mock_read_holding(*args, **kw):
-        call_n["n"] += 1
-        if call_n["n"] == 1:
-            return None  # batch fails
-        # probe attempt: 3-arg form raises TypeError first time
-        if call_n["n"] == 2:
-            raise TypeError("bad sig")
-        return [1]  # fallback 2-arg succeeds
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", side_effect=mock_read_holding),
-        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert "fake_holding" in result["available_registers"]["holding_registers"]
-
-
-# ---------------------------------------------------------------------------
-# Line 1634: scan_device legacy path returns non-tuple dict
-# ---------------------------------------------------------------------------
 
 
 @pytest.mark.asyncio
@@ -2410,72 +1683,6 @@ async def test_verify_connection_rtu_with_serial_port():
 
 # ---------------------------------------------------------------------------
 # Lines 1271, 1275: full_register_scan coil alias path and unknown address
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_coil_alias_path():
-    """Line 1271: full_register_scan coil with alias names updates all aliases."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    # addr=0 → "fake_coil" + alias "fake_coil_alias"
-    scanner._registers = {4: {}, 3: {}, 1: {0: "fake_coil"}, 2: {}}
-    scanner._names_by_address[1] = {0: {"fake_coil", "fake_coil_alias"}}
-
-    async def coil_mock(*args, **kw):
-        count = 1
-        for a in reversed(args):
-            if isinstance(a, int):
-                count = a
-                break
-        return [True] * count
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", side_effect=coil_mock),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    assert "fake_coil_alias" in result["available_registers"]["coil_registers"]
-
-
-@pytest.mark.asyncio
-async def test_full_register_scan_coil_unknown_addr():
-    """Line 1275: full_register_scan coil with address not in _registers[1]."""
-    scanner = await _make_scanner(full_register_scan=True, retry=1)
-    scanner._client = AsyncMock()
-    # coil_max=2, but only addr=2 is registered; addrs 0,1 are unknown
-    scanner._registers = {4: {}, 3: {}, 1: {2: "fake_coil"}, 2: {}}
-    scanner._names_by_address[1] = {2: {"fake_coil"}}
-
-    async def coil_mock(*args, **kw):
-        count = 1
-        for a in reversed(args):
-            if isinstance(a, int):
-                count = a
-                break
-        return [True] * count
-
-    with (
-        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
-        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
-        patch.object(scanner, "_read_coil", side_effect=coil_mock),
-        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
-    ):
-        result = await scanner.scan()
-
-    # addrs 0 and 1 should be in unknown_registers
-    assert 0 in result["unknown_registers"]["coil_registers"]
-
-
-# ---------------------------------------------------------------------------
-# Line 1296: full_register_scan discrete unknown address
 # ---------------------------------------------------------------------------
